// src/lib/bindings/GeneratedConnectorResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GeneratedToolDefinition } from "./GeneratedToolDefinition";

export type GeneratedConnectorResult = {
  connectorId: string;
  connectorName: string;
  connectorLabel: string;
  tools: Array<GeneratedToolDefinition>;
  credentialFields: Array<unknown>;
  healthcheckEndpoint: string | null;
};